refactor: pull out more option computation to options.py
